Fermat's Last Theorem
This program proves that Fermat's Last Theorem is correct which states that x to the power of n + y to the power of n = z to the power of n has no positive whole-number solutions when n is greater than 2. However, the scope of the program does not cover Andrew Wiles proof of this theory (algebraic geometry, number theory, modularity theorem, etc.).