Greatest Common Factor using recursion
This program is using the Euclidian algorithm in finding the greatest common factor of a number.
For example: a = 126 and b = 96
Firstly, the left side of expression should be a = b * quotient + remainder
128 = 96 * 1 + 30 (How many instances does right side will become equal to the left side.)
Secondly, the left side is destroyed and replaced by the current value of b. so a = b.
b -> a such that:
96 = 30 * 3 remainder 6
Subsequently, the left side is always destroyed and replaced by the value of b.
30 = 6 * 5 remainder 0
6 = 0
for b that will be a (b1 *b2), always take the biggest factor from the 2 number.
when the right reaches 0, the one on the left is the GCF of the 2 numbers.
Hence, if b==0, returns a.